feat: modularize drawing api
