+++ src/bearzzzbot/main.py
@@ -1,12 +1,10 @@
 from fastapi import Depends, FastAPI, Request
 from fastapi.responses import HTMLResponse
 from fastapi.templating import Jinja2Templates
-from sqlalchemy.orm import Session
 
 from .config import Settings, get_settings
 from .db import close_db_connection, connect_to_db
-from .deps import get_tx_session
-from .insert_db import insert_db
+from .routes import api_router
 
 
 def make_app() -> FastAPI:
@@ -24,22 +22,7 @@ def make_app() -> FastAPI:
             },
         )
 
-    @app.get("/data/{statement}")
-    async def send_data_to_db(
-        statement: str,
-        db_session: Session = Depends(get_tx_session),
-    ):
-        stm = statement.split("&")
-        data = []
-        data.append(stm[0])
-
-        for i in range(1, 6):
-            if stm[i] == "true":
-                data.append(1)
-            else:
-                data.append(0)
-        insert_db(db_session, data)
-        return "OK"
+    app.include_router(api_router, prefix="/api/v1")
 
     @app.on_event("startup")
     async def on_startup() -> None:  # pragma: no cover
